refactor(attempt-review): Decompose and sanitize rich text
