+++ package.json
@@ -29,6 +29,7 @@
     "axios": "^1.13.4",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "dompurify": "^3.4.8",
     "immutable": "^5.1.6",
     "lucide-react": "^0.563.0",
     "next-themes": "^0.4.6",
