+++ src/course/attempt-review/rich-text-content.component.tsx
@@ -1,0 +1,28 @@
+import { cn } from '@/shadcn/lib/utils';
+import { sanitizeRichTextHtml } from './rich-text-sanitize.dom';
+
+export function RichTextContent({
+  html,
+  className,
+}: {
+  html: string;
+  className?: string;
+}) {
+  if (!html) {
+    return (
+      <p className={cn('text-sm text-muted-foreground', className)}>
+        Нет сохранённого текста.
+      </p>
+    );
+  }
+
+  return (
+    <div
+      className={cn(
+        'attempt-review-editor-readonly prose prose-sm max-w-none dark:prose-invert',
+        className
+      )}
+      dangerouslySetInnerHTML={{ __html: sanitizeRichTextHtml(html) }}
+    />
+  );
+}
