+++ src/course/attempt-review/rich-text-sanitize.dom.ts
@@ -1,0 +1,34 @@
+import DOMPurify from 'dompurify';
+
+const ALLOWED_TAGS = [
+  'a',
+  'blockquote',
+  'br',
+  'code',
+  'em',
+  'h1',
+  'h2',
+  'h3',
+  'h4',
+  'h5',
+  'h6',
+  'hr',
+  'img',
+  'li',
+  'ol',
+  'p',
+  'pre',
+  's',
+  'strong',
+  'ul',
+];
+
+const ALLOWED_ATTR = ['alt', 'href', 'rel', 'src', 'target', 'title'];
+
+export function sanitizeRichTextHtml(html: string): string {
+  return DOMPurify.sanitize(html, {
+    ALLOWED_ATTR,
+    ALLOWED_TAGS,
+    ALLOWED_URI_REGEXP: /^(?:(?:https?|mailto|tel):|\/(?!\/)|#)/i,
+  });
+}
